Hw7: Added save and load functionalities to the game. 1. Synchronously playing: After the player logged in, the player can click the Join button to enter the auto-matching process. If at the same period another player comes in. The server will match the two players and the game will start immediately.
2. Asynchronously playing: The player can click the start button to
start a new game with another player by specifying his or her email
address. The chosen opponent has to already logged in once into the game
to be found. After the game started, the player started the game can
make a move immediately (Currently the player who start the game will be
assigned to be the BLACK player). 

3. The Game List button is for the players to switch between on-going
games. Any available game will show up with the status of the game (Game
id, whose turn it is etc.) in a pop-up when the player click the button.

AppEngine Link: http://5.weiqi-hezzze.appspot.com